[backend] chore: diary 도메인에서 사용되는 모든 ID 타입을 int → long 으로 수정

+++ backend/src/main/java/com/example/demo/mapper/DiaryMapper.java
@@ -18,11 +18,11 @@ public interface DiaryMapper {
 
     void updateDiary(DiaryEditRequestDTO diary); // 일기 수정
 
-    void deleteDiary(int diaryId);  // 일기 삭제
+    void deleteDiary(long diaryId);  // 일기 삭제
 
-    List<AllTeamDiariesResponse> requestAllTeamDiaries(int userId); // 사용자가 속해있는 모든 그룹의 일기 조회
+    List<AllTeamDiariesResponse> requestAllTeamDiaries(long userId); // 사용자가 속해있는 모든 그룹의 일기 조회
 
     DiaryModel requestDiaryDetails(long diaryId);    // 일기 상세 내용 요청
 
-    List<DiaryModel> requestDiaryId(String diaryTitle, String writtenDate, int writerId);    // 일기 아이디 요청
+    List<DiaryModel> requestDiaryId(String diaryTitle, String writtenDate, long writerId);    // 일기 아이디 요청
 }
