+++ backend/src/main/java/com/example/demo/repository/DiaryRepository.java
@@ -14,7 +14,7 @@ public interface DiaryRepository {
     void updateDiary(DiaryEditRequestDTO diaryEditRequestDTO);
 
     DiaryDetailResponseDTO requestDiaryDetails(long diaryId);
-    void deleteDiary(int diaryId);
-    List<AllTeamDiariesResponse> requestAllTeamDiaries(int userId);
-    List<DiaryModel> requestDiaryId(String diaryTitle, String writtenDate, int writerId);
+    void deleteDiary(long diaryId);
+    List<AllTeamDiariesResponse> requestAllTeamDiaries(long userId);
+    List<DiaryModel> requestDiaryId(String diaryTitle, String writtenDate, long writerId);
 }
